+++ crates/test_utils/wasm/src/lib.rs
@@ -50,6 +50,8 @@ pub enum TestWasm {
     Link,
     MigrateAgentFail,
     MigrateAgentPass,
+    MigrateInitial,
+    MigrateNew,
     MultipleCalls,
     MustGet,
     MustGetAgentActivity,
@@ -159,6 +161,8 @@ impl From<TestWasm> for ZomeName {
             TestWasm::Link => "link",
             TestWasm::MigrateAgentFail => "migrate_agent_fail",
             TestWasm::MigrateAgentPass => "migrate_agent_pass",
+            TestWasm::MigrateInitial => "migrate_initial",
+            TestWasm::MigrateNew => "migrate_new",
             TestWasm::MultipleCalls => "multiple_calls",
             TestWasm::MustGet => "must_get",
             TestWasm::MustGetAgentActivity => "must_get_agent_activity",
@@ -247,6 +251,12 @@ impl From<TestWasm> for PathBuf {
             TestWasm::MigrateAgentPass => {
                 "wasm32-unknown-unknown/release/test_wasm_migrate_agent_pass.wasm"
             }
+            TestWasm::MigrateInitial => {
+                "wasm32-unknown-unknown/release/test_wasm_migrate_initial.wasm"
+            }
+            TestWasm::MigrateNew => {
+                "wasm32-unknown-unknown/release/test_wasm_migrate_new.wasm"
+            },
             TestWasm::MultipleCalls => {
                 "wasm32-unknown-unknown/release/test_wasm_multiple_calls.wasm"
             }